moving template to e2e test folder
