--- deployments/e2e-network/_template-1/_infrastructure/aliases.json
@@ -1,26 +1,0 @@
-{
-    "fauceteer": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
-    "governor": "0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9",
-    "governor:implementation": "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
-    "COMP": "0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0",
-    "timelock": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512",
-    "cometAdmin": "0x0165878A594ca255338adfa4d48449f69242Eb8F",
-    "cometFactory": "0x8A791620dd6260079BF849Dc5567aDC3F2FdC318",
-    "configurator": "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6",
-    "configurator:implementation": "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853",
-    "configuratorAdmin": "0x0165878A594ca255338adfa4d48449f69242Eb8F",
-    "rewards": "0x610178dA211FEF7D417bC0e6FeD39F05609AD788",
-    "DAI": "0xA51c1fc2f0D1a1b8494Ed1FE312d7C3a78Ed91C0",
-    "USDC": "0x0DCd1Bf9A1b36cE34237eEaFef220932846BCD82",
-    "WETH": "0x9A676e781A523b5d0C0e43731313A708CB607508",
-    "WBTC": "0x0B306BF915C4d645ff596e518fAf3F9669b97016",
-    "LINK": "0x959922bE3CAee4b8Cd9a407cc3ac1C251C2007B1",
-    "UNI": "0x9A9f2CCfdE556A7E9Ff0848998Aa4a0CFD8863AE",
-    "daiPriceFeed": "0x68B1D87F95878fE05B998F19b66F4baba5De1aed",
-    "usdcPriceFeed": "0x3Aa5ebB10DC797CAC828524e59A333d0A371443c",
-    "wethPriceFeed": "0xc6e7DF5E7b4f2A278906862b61205850344D4e7d",
-    "wbtcPriceFeed": "0x59b670e9fA9D0A427751Af201D676719a970857b",
-    "compPriceFeed": "0x4ed7c70F96B99c776995fB64377f0d4aB3B0e1C1",
-    "linkPriceFeed": "0x322813Fd9A801c5507c9de605d63CEA4f2CE6c44",
-    "uniPriceFeed": "0xa85233C63b9Ee964Add6F2cffe00Fd84eb32338f"
-}
