+++ test/e2e/_template-1/dai/configuration.json
