+++ test/e2e/_template-1/dai/deploy.ts
