+++ test/e2e/_template-1/usdc/deploy.ts
