+++ test/e2e/_template-1/usdc/relations.ts
